Prototype
This draft website is a carefully constructed mockup, intended to refine my skills and bolster my motivation. It stands as a testament to my self-taught journey, providing a practical framework for learning and growth. This preliminary draft is a personal reminder of my potential, a virtual canvas on which I practice and perfect my craft.